refactor: Move external service under adapter
