+++ backend_rust/src/adapters.rs
@@ -1,3 +1,3 @@
 pub mod controller;
-pub mod grpc;
 pub mod repository;
+pub mod external;
