+++ backend_rust/src/adapters/external/grpc.rs
